business-server/pyproject.toml
[project]
name = "business-server"
version = "0.1.0"
description = "AI Powered Hardware Store Management System"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.128.0",
    "fastapi-users[sqlalchemy]>=15.0.3",
    "httpx>=0.28.1",
    "imagekitio==4.2.0",
    "psycopg2>=2.9.11",
    "pydantic-settings>=2.12.0",
    "aiosqlite>=0.20.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.15.1",
    "python-jose>=3.5.0",
    "stripe>=14.3.0",
    "uvicorn>=0.40.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
